+++ src/App.tsx
@@ -1,17 +1,17 @@
 import './App.css';
 import { RecoilRoot } from 'recoil';
 import { Routes, Route, BrowserRouter } from 'react-router-dom';
-import Home from 'pages/Home';
+import Home from 'pages/home/Home';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import DetailPetPage from 'pages/detailPet/DetailPetPage';
-import ProfileListPage from 'pages/ProfileList/ProfileListPage';
-import LoginPage from 'pages/LoginPage';
+import ProfileListPage from 'pages/profileList/ProfileListPage';
+import LoginPage from 'pages/login/LoginPage';
 import MapPage from 'pages/map/MapPage';
-import NewListPage from 'pages/ProfileList/newList/NewListPage';
-import RegisterPage from 'pages/RegisterPage';
-import ShelterInfoPage from 'pages/ShelterInfoPage';
-import SignupPage from 'pages/SignupPage';
-import UrgentListPage from 'pages/ProfileList/urgentList/UrgentListPage';
+import NewListPage from 'pages/profileList/newList/NewListPage';
+import RegisterPage from 'pages/register/RegisterPage';
+import ShelterInfoPage from 'pages/shelterInfo/ShelterInfoPage';
+import SignupPage from 'pages/signUp/SignupPage';
+import UrgentListPage from 'pages/profileList/urgentList/UrgentListPage';
 
 const queryClient = new QueryClient();
